# chruch-cms/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  version: 1.0.0
  description: >
    REST API for a web-based Church Management System (CMS).
    Auth is handled by Supabase Auth. This API verifies Supabase JWTs and enforces RBAC.
servers:
  - url: http://localhost:8081
    description: Local dev server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /api/members
      required: [timestamp, status, error, message, path]
    Role:
      type: string
      enum: [ADMIN, LEADER, MEMBER]
...
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
          example: John Banda
        email:
          type: string
          format: email
          example: john.banda@example.com
        phone:
          type: string
          example: "+26097XXXXXXX"
        role:
          $ref: '#/components/schemas/Role'
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          example: ACTIVE
        joinDate:
          type: string
          format: date
          example: "2025-01-12"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, fullName, role, status, joinDate]
    MemberCreateRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        joinDate:
          type: string
          format: date
      required: [fullName, role, status, joinDate]
    MemberUpdateRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        joinDate:
          type: string
          format: date
    PagedMembersResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 125
        totalPages:
          type: integer
          example: 7
      required: [items, page, size, totalItems, totalPages]
...
security:
  - BearerAuth: []
paths:
...
  /api/members:
    get:
      summary: List members (Admin only)
      tags: [Members]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
        - in: query
          name: q
          description: Optional search (name/email/phone)
          schema:
            type: string
      responses:
        '200':
          description: Paged members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMembersResponse'
        '403':
          description: Forbidden (role not allowed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create member (Admin only)
      tags: [Members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreateRequest'
      responses:
        '201':
          description: Member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/members/{memberId}:
    get:
      summary: Get member by ID (Admin only)
      tags: [Members]
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update member (Admin only)
      tags: [Members]
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdateRequest'
      responses:
        '200':
          description: Updated member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete member (Admin only)
      tags: [Members]
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...